+++ package.json
@@ -6,6 +6,10 @@
     ".": "./src/index.js",
     "./package.json": "./package.json"
   },
+  "files": [
+    "./src/*",
+    "README.md"
+  ],
   "description": "A DocumentPersistentFragment alternative, solely based on a weak relation between a fragment and its comments' boundaries",
   "main": "./src/index.js",
   "module": "./src/index.js",
